# [pkg/ottl] Improve time performance