@atoulme 2024-09-10
I fed your benchmark results into benchstat:
```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ottl/ottlfuncs
                                                                        │  /tmp/before.txt   │               /tmp/after.txt                │
                                                                        │       sec/op       │       sec/op        vs base                 │
_Time/simple_short_form-10                                                0.000007900n ± ∞ ¹   0.000005400n ± ∞ ¹        ~ (p=1.000 n=1) ²
_Time/simple_short_form_with_short_year_and_slashes-10                    0.000011500n ± ∞ ¹   0.000003700n ± ∞ ¹        ~ (p=1.000 n=1) ²
_Time/month_day_year-10                                                   0.000005700n ± ∞ ¹   0.000005300n ± ∞ ¹        ~ (p=1.000 n=1) ²
_Time/simple_long_form-10                                                 0.000007500n ± ∞ ¹   0.000004200n ± ∞ ¹        ~ (p=1.000 n=1) ²
_Time/date_with_timestamp-10                                              0.000006300n ± ∞ ¹   0.000008700n ± ∞ ¹        ~ (p=1.000 n=1) ²
_Time/day_of_the_week_long_form-10                                        0.000008500n ± ∞ ¹   0.000003500n ± ∞ ¹        ~ (p=1.000 n=1) ²
_Time/short_weekday,_short_month,_long_format-10                          0.000008900n ± ∞ ¹   0.000003600n ± ∞ ¹        ~ (p=1.000 n=1) ²
_Time/short_months-10                                                     0.000007000n ± ∞ ¹   0.000003100n ± ∞ ¹        ~ (p=1.000 n=1) ²
_Time/timestamp_with_time_zone_offset-10                                   0.00006650n ± ∞ ¹    0.00004910n ± ∞ ¹        ~ (p=1.000 n=1) ²
_Time/short_date_with_timestamp_without_time_zone_offset-10                0.00004280n ± ∞ ¹    0.00003810n ± ∞ ¹        ~ (p=1.000 n=1) ²
_Time/RFC_3339_in_custom_format-10                                         0.00003450n ± ∞ ¹    0.00003650n ± ∞ ¹        ~ (p=1.000 n=1) ²
_Time/RFC_3339_in_custom_format_before_2000-10                             0.00003490n ± ∞ ¹    0.00003640n ± ∞ ¹        ~ (p=1.000 n=1) ²
_Time/no_location-10                                                      0.000003500n ± ∞ ¹   0.000002800n ± ∞ ¹        ~ (p=1.000 n=1) ²
_Time/with_location_-_America-10                                          0.000010400n ± ∞ ¹   0.000001700n ± ∞ ¹        ~ (p=1.000 n=1) ²
_Time/with_location_-_Asia-10                                             0.000008400n ± ∞ ¹   0.000002800n ± ∞ ¹        ~ (p=1.000 n=1) ²
_Time/RFC_3339_in_custom_format_before_2000,_ignore_default_location-10    0.00003790n ± ∞ ¹    0.00003930n ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                                                    0.00001279n         0.000007823n        -38.85%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```

38.85% performance gains!